--- area-backend/src/common/interfaces/fields.type.ts
@@ -1,5 +1,0 @@
-export interface Field {
-    name: string;
-    type: 'string' | 'number' | 'boolean';
-    required: boolean;
-}
