--- area-backend/src/common/interfaces/reactions.type.ts
@@ -1,6 +1,0 @@
-import { Field } from "./fields.type";
-
-export interface Reactions {
-    run(userId: string, params: any): Promise<void>;
-    getFields(): [Field]
-}
