+++ area-backend/src/modules/auth/core/provider-registry-impl.ts
@@ -1,36 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { JwtService } from '@nestjs/jwt';
-import { ConfigService } from '@nestjs/config';
-import { IdentityProvider, LinkingProvider, ProviderKey } from './OAuth2Types';
-import { TokenCrypto } from './TokenCrypto';
-import { TokenStore } from './TokenStore';
-
-/**
- * Dependency set available to provider plugins (config, jwt, crypto, token store).
- */
-export interface PluginDeps {
-  config: ConfigService;
-  jwt: JwtService;
-  crypto: TokenCrypto;
-  store: TokenStore;
-}
-
-/**
- * Read-only registry to resolve identity and linking providers by key.
- */
-export interface ProviderRegistry {
-  getIdentity(provider: ProviderKey): IdentityProvider | undefined;
-  getLinking(provider: ProviderKey): LinkingProvider | undefined;
-  listProviders(): ProviderKey[];
-}
-
-/**
- * Mutable builder for assembling the provider registry.
- */
-export interface ProviderRegistryBuilder extends ProviderRegistry {
-  addIdentity(p: IdentityProvider): void;
-  addLinking(p: LinkingProvider): void;
-}
+import { IdentityProvider, LinkingProvider, ProviderKey } from '../../../common/interfaces/oauth2.type';
+import type { ProviderRegistry, ProviderRegistryBuilder } from '../../../common/interfaces/provider-registry.type';
 
 @Injectable()
 export class ProviderRegistryImpl implements ProviderRegistryBuilder {
