+++ area-backend/src/modules/auth/providers/oauth2/spotify.link.provider.ts
@@ -1,9 +1,10 @@
-import { OAuth2LinkProvider } from './OAuth2LinkProvider';
+import { OAuth2LinkProvider } from './link.provider';
 import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
-import { UsersService, ProviderKey } from '../../../users/users.service';
+import { UsersService } from '../../../users/users.service';
+import { ProviderKeyEnum as ProviderKey } from 'src/common/interfaces/oauth2.type';
 import { UnauthorizedException, InternalServerErrorException, BadRequestException, Logger } from '@nestjs/common';
-import { TokenCryptoUtil } from '../TokenCryptoUtil';
+import { AesGcmTokenCrypto } from '../../core/token-crypto';
 
 /**
  * Spotify OAuth linker that connects a Spotify account to an existing user.
@@ -17,7 +18,7 @@ export class SpotifyOAuthLinkProvider implements OAuth2LinkProvider {
     private readonly usersService: UsersService,
     private readonly config: ConfigService,
     private readonly jwtService: JwtService,
-    private readonly crypto: TokenCryptoUtil,
+    private readonly crypto: AesGcmTokenCrypto,
   ) {}
 
   /**
