+++ area-backend/src/modules/manager/manager.controller.ts
@@ -1,7 +1,8 @@
 import { Controller, Get, Param, Post, Body, Delete, HttpCode, HttpStatus } from '@nestjs/common';
-import { ManagerService, ActionCallback, ReactionCallback } from './manager.service';
+import { ManagerService } from './manager.service';
+import { ActionCallback, ReactionCallback } from '../../common/interfaces/area.type';
 import { ApiBody } from '@nestjs/swagger/dist/decorators/api-body.decorator';
-import { ApiParam, ApiProperty } from '@nestjs/swagger';
+import { ApiOperation, ApiParam, ApiProperty } from '@nestjs/swagger';
 
 class BindActionDto {
   @ApiProperty({ description: 'Name of the action to bind', example: 'new_email' })
@@ -22,6 +23,7 @@ export class ManagerController {
    * @returns List of registered action callbacks
    */
   @Get('actions')
+  @ApiOperation({ summary: 'Get available actions' })
   getAvailableActions(): ActionCallback[] {
     return this.managerService.getAvailableActions();
   }
@@ -31,6 +33,7 @@ export class ManagerController {
    * @returns List of registered reaction callbacks
    */
   @Get('reactions')
+  @ApiOperation({ summary: 'Get available reactions' })
   getAvailableReactions(): ReactionCallback[] {
     return this.managerService.getAvailableReactions();
   }
@@ -44,6 +47,7 @@ export class ManagerController {
   @Post('areas/:userId')
   @ApiBody({ type: BindActionDto })
   @ApiParam({ name: 'userId', description: 'ID of the user' })
+  @ApiOperation({ summary: 'Bind an action to a reaction for a user' })
   async bindAction(@Param('userId') userId: string, @Body() bindActionDto: BindActionDto) {
     const areaId = await this.managerService.bindAction(
       userId,
@@ -64,6 +68,8 @@ export class ManagerController {
    * @param userId - Target user ID
    */
   @Get('areas/:userId')
+  @ApiParam({ name: 'userId', description: 'ID of the user' })
+  @ApiOperation({ summary: 'Get all areas for a user' })
   async getUserAreas(@Param('userId') userId: string) {
     return await this.managerService.getUserAreas(userId);
   }
@@ -74,6 +80,8 @@ export class ManagerController {
    */
   @Delete('areas/:areaId')
   @HttpCode(HttpStatus.NO_CONTENT)
+  @ApiParam({ name: 'areaId', description: 'ID of the area to deactivate' })
+  @ApiOperation({ summary: 'Deactivate an area' })
   async deactivateArea(
     @Param('areaId') areaId: string
   ) {
